pyproject.toml
[project]
name = "markkurzeja-github-io"
version = "0.1.0"
description = "TensorTales - personal blog"
requires-python = ">=3.13"
dependencies = [
    "markdown>=3.10.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100
indent-width = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
